Add custom method for horizontally centering arrow symbols
This is because using a normal, rectangular centering looks visually wrong for this shape, switching to a 'center of mass' centering looks better and was enough to persuade me to make this tiny custom tweak. Will add one for up and down triangles at some point too so it can be added to the default scroll bar theming.