Add new allowed char groups and password style text hiding
This is for the UITextLine element and came form user requests. The password box style display can be toggle on or off and disables users using CTRL+C/CTRL+X to copy the password from the box into plain text. The actual text typed in is still stored in the UI element and is accessible by code, only the display is affected.