Changed the variable which denotes the valid pygame_gui named colours from __colourNames__ to _namedColours to comply with the PEP 8 Style Guide: “__double_leading_and_trailing_underscore__: “magic” objects or attributes that live in user-controlled namespaces. E.g. __init__, __import__ or __file__. Never invent such names; only use them as documented.” Additionally, __colourNames__ implies a list of colour names, while _namedColours implies more toward the actual purpose of the dictionary
Added convention where colour name strings (e.g. “red”, “green”, “black”, etc…) are not case-sensitive.
	Dev Notes:
This additionally means that if any new colours are added to the _namedColours dictionary in pygame_gui._constants.py, they must be lowercase in order to properly work with pygame_gui. There’s a test in test_constants.py added to detect this if it ever happens however, so it shouldn’t go unnoticed

Made colours to the colour format table in the Theme Guide page of the pygame_gui documentation actually display which colour they are describing

Added the “Colour Name” row to the colour format table in the Theme Guide page of the pygame_gui documentation for transparency on its availability as a colour theming option

Changed the “Colour: “ warning in pygame_gui.core.ui_appearance_theme.py when parsing colour data to an “Invalid Theme Colour: “ warning, needed a more descriptive warning header

Added warning suppressions for the “Invalid Theme Colour: “ header to respective unit tests

Finished the “Parsing a Gradient: “ section of documentation on the module docstring of pygame_gui.core.colour_parser.py