Update layout_guide.rst
Punctuation and grammatical fixes. It's = contraction of "it is". Its = possessive. The name 'pygame' (outside of code) is actually PyGame (the official capitalization). Some sentences did not make grammatical sense due to improper placement, or lack of, commas. Changed "interrogate" to "utilize" since that would have the sentence make more sense.