Fix #610 by not hiding if not visible
This fixes issue #610, where calling the hide() method on an already hidden element called the full hide() functionality again, including generating any events, such as unhovering. This bug is present in all hide()  method implementation in the codebase (except UIPanel.hide) and is fixed with this commit for all widgets by exiting the hide() method early if self.visible is False.